bugfix the problem that multiple config items may overflow

+++ include/pika_conf.h
@@ -42,7 +42,7 @@ class PikaConf : public slash::BaseConf {
   int db_sync_speed()   { RWLock l(&rwlock_, false); return db_sync_speed_; }
   std::string compact_cron() { RWLock l(&rwlock_, false); return compact_cron_; }
   std::string compact_interval() { RWLock l(&rwlock_, false); return compact_interval_; }
-  int write_buffer_size() { RWLock l(&rwlock_, false); return write_buffer_size_; }
+  int64_t write_buffer_size() { RWLock l(&rwlock_, false); return write_buffer_size_; }
   int timeout()           { RWLock l(&rwlock_, false); return timeout_; }
   std::string server_id() { RWLock l(&rwlock_, false); return server_id_; }
 
@@ -66,8 +66,8 @@ class PikaConf : public slash::BaseConf {
   int max_background_compactions()   { RWLock l(&rwlock_, false); return max_background_compactions_; }
   int max_cache_files()          { RWLock l(&rwlock_, false); return max_cache_files_; }
   int max_bytes_for_level_multiplier() {RWLock l(&rwlock_, false); return max_bytes_for_level_multiplier_; }
-  int block_size() {RWLock l(&rwlock_, false); return block_size_; }
-  int block_cache() {RWLock l(&rwlock_, false); return block_cache_; }
+  int64_t block_size() {RWLock l(&rwlock_, false); return block_size_; }
+  int64_t block_cache() {RWLock l(&rwlock_, false); return block_cache_; }
   bool share_block_cache() {RWLock l(&rwlock_, false); return share_block_cache_; }
   bool cache_index_and_filter_blocks() {RWLock l(&rwlock_, false); return cache_index_and_filter_blocks_; }
   bool optimize_filters_for_hits() {RWLock l(&rwlock_, false); return optimize_filters_for_hits_; }
@@ -208,7 +208,7 @@ private:
   int db_sync_speed_;
   std::string compact_cron_;
   std::string compact_interval_;
-  int write_buffer_size_;
+  int64_t write_buffer_size_;
   int log_level_;
   bool daemonize_;
   bool slotmigrate_;
@@ -238,8 +238,8 @@ private:
   int max_background_compactions_;
   int max_cache_files_;
   int max_bytes_for_level_multiplier_;
-  int block_size_;
-  int block_cache_;
+  int64_t block_size_;
+  int64_t block_cache_;
   bool share_block_cache_;
   bool cache_index_and_filter_blocks_;
   bool optimize_filters_for_hits_;
